update Access model

+++ amember/model.go
@@ -127,11 +127,11 @@ type Access struct {
 	ProductTitle string `json:"product_title"`
 	Status bool `json:"status"`
 	ProductDescription string `json:"product_description"`
-	Spend float64 `json:"spend"`
-	SpendCoveredByPlan float64 `json:"spend_covered_by_plan"`
-	Overage float64 `json:"overage"`
-	ProjectedSpend float64 `json:"projected_spend"`
-	ProjectedOverage float64 `json:"projected_overage"`
+	Spend float32 `json:"spend"`
+	SpendCoveredByPlan float32 `json:"spend_covered_by_plan"`
+	Overage float32 `json:"overage"`
+	ProjectedSpend float32 `json:"projected_spend"`
+	ProjectedOverage float32 `json:"projected_overage"`
 	
 }
